Complete function and associated helper functions for markdown_to_html_node. Encompasses a reorg of modules into helpers.py in order to address  circular import issues.
