+++ src/htmlnode.py
@@ -1,16 +1,3 @@
-from constants import (
-    TAG_TYPE_HEADING,
-    TAG_TYPE_LIST_ITEM,
-    TAG_TYPE_UNORDERED_LIST,
-    TAG_TYPE_ORDERED_LIST,
-    TAG_TYPE_DIV,
-)
-
-from constants import BLOCK_TYPE_QUOTE
-
-from textnode import markdown_to_blocks, block_to_block_type, text_to_textnodes
-
-
 class HTMLNode:
     def __init__(self, tag=None, value=None, children=None, props=None):
         self.tag = tag
@@ -69,46 +56,3 @@ class ParentNode(HTMLNode):
         for child in self.children:
             output += child.to_html()
         return f"{output}</{self.tag}>"
-
-
-def blockquote_to_html_node(text):
-    return LeafNode("blockquote", text)
-
-
-def paragraph_to_html_node(text):
-    return LeafNode("p", text)
-
-
-def code_to_html_node(text):
-    child_node = LeafNode("code", text)
-    return ParentNode("pre", [child_node])
-
-
-def heading_to_html_node(text):
-    tag = TAG_TYPE_HEADING + str(text.count("#"))
-    return LeafNode(tag, text)
-
-
-def unordered_list_to_html_node(text):
-    child_list_items = []
-    for list_item in text.split("\n"):
-        list_item = list_item[1:]
-        list_item = list_item.strip()
-        child_list_items.append(LeafNode(TAG_TYPE_LIST_ITEM, list_item))
-    return ParentNode(TAG_TYPE_UNORDERED_LIST, child_list_items)
-
-
-def ordered_list_to_html_node(text):
-    child_list_items = []
-    for list_item in text.split("\n"):
-        list_item = list_item.strip()
-        child_list_items.append(LeafNode(TAG_TYPE_LIST_ITEM, list_item))
-    return ParentNode(TAG_TYPE_ORDERED_LIST, child_list_items)
-
-def markdown_to_html_node(markdown):
-    blocks = markdown_to_blocks(markdown)
-    child_nodes = []
-    for block in blocks:
-        if block_to_block_type(block) == BLOCK_TYPE_QUOTE:
-            text_nodes = text_to_textnodes(block)
-    return ParentNode(TAG_TYPE_DIV, child_nodes)
