+++ src/test_textnode.py
@@ -1,23 +1,5 @@
 import unittest
-from textnode import (
-    TextNode,
-    split_nodes_delimiter,
-    extract_markdown_images,
-    extract_markdown_links,
-    split_nodes_image,
-    split_nodes_link,
-    text_to_textnodes,
-    markdown_to_blocks,
-    block_to_block_type,
-)
-from constants import (
-    BLOCK_TYPE_ORDERED_LIST,
-    BLOCK_TYPE_HEADING,
-    BLOCK_TYPE_PARAGRAPH,
-    BLOCK_TYPE_CODE,
-    BLOCK_TYPE_QUOTE,
-    BLOCK_TYPE_UNORDERED_LIST,
-)
+from textnode import TextNode
 
 
 class TestTextNode(unittest.TestCase):
@@ -77,286 +59,6 @@ class TestTextNode(unittest.TestCase):
         self.assertEqual(leaf_node.tag, "img")
         self.assertEqual(leaf_node.value, "")
         self.assertEqual(leaf_node.props, {"src": "https://example.com", "alt": ""})
-
-
-class TestSplitNodesDelimiter(unittest.TestCase):
-    def test_split_nodes_delimiter_code(self):
-        node = TextNode("This is text with a `code block` word", "text")
-        self.assertEqual(
-            split_nodes_delimiter([node], "`", "code"),
-            [
-                TextNode("This is text with a ", "text"),
-                TextNode("code block", "code"),
-                TextNode(" word", "text"),
-            ],
-        )
-
-    def test_split_nodes_delimiter_bold(self):
-        node = TextNode("This is text with a *bold* word", "text")
-        self.assertEqual(
-            split_nodes_delimiter([node], "*", "bold"),
-            [
-                TextNode("This is text with a ", "text"),
-                TextNode("bold", "bold"),
-                TextNode(" word", "text"),
-            ],
-        )
-
-    def test_split_nodes_delimiter_italic(self):
-        node = TextNode("This is text with a **italic** word", "text")
-        self.assertEqual(
-            split_nodes_delimiter([node], "**", "italic"),
-            [
-                TextNode("This is text with a ", "text"),
-                TextNode("italic", "italic"),
-                TextNode(" word", "text"),
-            ],
-        )
-
-    def test_split_nodes_delimiter_non_text_type(self):
-        nodes = [
-            TextNode("This is text with a *bold* word", "text"),
-            TextNode("This is a non text type node", "code"),
-        ]
-        self.assertEqual(
-            split_nodes_delimiter(nodes, "*", "bold"),
-            [
-                TextNode("This is text with a ", "text"),
-                TextNode("bold", "bold"),
-                TextNode(" word", "text"),
-                TextNode("This is a non text type node", "code"),
-            ],
-        )
-
-
-class TestExtractMarkdown(unittest.TestCase):
-    def test_extract_markdown_images(self):
-        text = "This is text with an ![image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and ![another](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/dfsdkjfd.png)"
-        self.assertEqual(
-            extract_markdown_images(text),
-            [
-                (
-                    "image",
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                (
-                    "another",
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/dfsdkjfd.png",
-                ),
-            ],
-        )
-
-    def test_extract_markdown_links(self):
-        text = "This is text with a [link](https://www.example.com) and [another](https://www.example.com/another)"
-        self.assertEqual(
-            extract_markdown_links(text),
-            [
-                ("link", "https://www.example.com"),
-                ("another", "https://www.example.com/another"),
-            ],
-        )
-
-
-class TestSplitNodesImage(unittest.TestCase):
-    # need to test multiple nodes passed in
-    def test_split_nodes_image_multiple(self):
-        node = TextNode(
-            "This is text with an ![image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and another ![second image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/3elNhQu.png)",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_image([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with an ", TextNode.text_type_text),
-                TextNode(
-                    "image",
-                    TextNode.text_type_image,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                TextNode(" and another ", TextNode.text_type_text),
-                TextNode(
-                    "second image",
-                    TextNode.text_type_image,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/3elNhQu.png",
-                ),
-            ],
-        )
-
-    def test_split_nodes_image_single(self):
-        node = TextNode(
-            "This is text with an ![image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png)",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_image([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with an ", TextNode.text_type_text),
-                TextNode(
-                    "image",
-                    TextNode.text_type_image,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-            ],
-        )
-
-    def test_split_nodes_image_single_with_ending_node(self):
-        node = TextNode(
-            "This is text with an ![image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and some additional text",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_image([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with an ", TextNode.text_type_text),
-                TextNode(
-                    "image",
-                    TextNode.text_type_image,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                TextNode(" and some additional text", TextNode.text_type_text),
-            ],
-        )
-
-
-class TestSplitNodesLink(unittest.TestCase):
-    def test_split_nodes_link_multiple(self):
-        node = TextNode(
-            "This is text with a [link](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and another [second link](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/3elNhQu.png)",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_link([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with a ", TextNode.text_type_text),
-                TextNode(
-                    "link",
-                    TextNode.text_type_link,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                TextNode(" and another ", TextNode.text_type_text),
-                TextNode(
-                    "second link",
-                    TextNode.text_type_link,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/3elNhQu.png",
-                ),
-            ],
-        )
-
-    def test_split_nodes_link_single(self):
-        node = TextNode(
-            "This is text with a [link](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png)",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_link([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with a ", TextNode.text_type_text),
-                TextNode(
-                    "link",
-                    TextNode.text_type_link,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-            ],
-        )
-
-    def test_split_nodes_link_single_with_ending_node(self):
-        node = TextNode(
-            "This is text with a [link](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and some additional text",
-            TextNode.text_type_text,
-        )
-        new_nodes = split_nodes_link([node])
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is text with a ", TextNode.text_type_text),
-                TextNode(
-                    "link",
-                    TextNode.text_type_link,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                TextNode(" and some additional text", TextNode.text_type_text),
-            ],
-        )
-
-
-class TestTextToTextNodes(unittest.TestCase):
-    def test_text_to_textnodes(self):
-        text = "This is **text** with an *italic* word and a `code block` and an ![image](https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png) and a [link](https://boot.dev)"
-        new_nodes = text_to_textnodes(text)
-        self.assertEqual(
-            new_nodes,
-            [
-                TextNode("This is ", TextNode.text_type_text),
-                TextNode("text", TextNode.text_type_bold),
-                TextNode(" with an ", TextNode.text_type_text),
-                TextNode("italic", TextNode.text_type_italic),
-                TextNode(" word and a ", TextNode.text_type_text),
-                TextNode("code block", TextNode.text_type_code),
-                TextNode(" and an ", TextNode.text_type_text),
-                TextNode(
-                    "image",
-                    TextNode.text_type_image,
-                    "https://storage.googleapis.com/qvault-webapp-dynamic-assets/course_assets/zjjcJKZ.png",
-                ),
-                TextNode(" and a ", TextNode.text_type_text),
-                TextNode("link", TextNode.text_type_link, "https://boot.dev"),
-            ],
-        )
-
-
-class TestMarkdownToBlocks(unittest.TestCase):
-    def test_markdown_to_blocks(self):
-        markdown_string = """
-            # This is a heading
-
-            This is a paragraph of text. It has some **bold** and *italic* words inside of it.
-
-            * This is a list item
-            * This is another list item
-        """
-        self.assertEqual(
-            markdown_to_blocks(markdown_string),
-            [
-                "# This is a heading",
-                "This is a paragraph of text. It has some **bold** and *italic* words inside of it.",
-                "* This is a list item\n            * This is another list item",
-            ],
-        )
-
-
-class TestBlockToBlockType(unittest.TestCase):
-    def test_block_to_block_type_heading(self):
-        block = "# This is a heading block"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_HEADING)
-
-    def test_block_to_block_type_code(self):
-        block = "```This is \n some \n code```"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_CODE)
-
-    def test_block_to_block_type_quote(self):
-        block = "> Every line in a quote block\n> has to start with a '> ' sequence"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_QUOTE)
-
-    def test_block_to_block_type_not_quote(self):
-        block = "> Every line in a quote block\n has to start with a '> ' sequence"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_PARAGRAPH)
-
-    def test_block_to_block_type_unordered_list(self):
-        block = "- This is \n* a list\n- with multiple types of\n* bullets"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_UNORDERED_LIST)
-
-    def test_block_to_block_type_ordered_list(self):
-        block = "1. This is \n2. an ordered list\n3. of multiple lines"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_ORDERED_LIST)
-
-    def test_block_to_block_type_not_ordered_list(self):
-        block = "1. This is \n3. an ordered list\n3. that is incorrectly formatted"
-        self.assertEqual(block_to_block_type(block), BLOCK_TYPE_PARAGRAPH)
 
 
 if __name__ == "__main__":
